* Refactored naming of classes for better understanding

+++ tutorials/DomainEventStore/src/main/java/io/jexxa/tutorials/domaineventstore/domain/valueobject/ContractNumber.java
@@ -11,7 +11,7 @@ public class ContractNumber
         this.value = value;
     }
 
-    public double getValue()
+    public int getValue()
     {
         return value;
     }
